% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_mean.R
\name{compute_mean}
\alias{compute_mean}
\title{Given a set of bigwig files, create a scaled mean BigWig file.}
\usage{
compute_mean(bws, outfile = "mean", aucs = NULL,
  chr_sizes = "/dcl01/lieber/ajaffe/Emily/RNAseq-pipeline/Annotation/hg38.chrom.sizes.gencode",
  auc_scale = 4e+07 * 100, wiggletools = "wiggletools",
  wigToBigWig = "wigToBigWig", tempdir = getwd())
}
\arguments{
\item{bws}{A named vector with the paths to the bigWig files. The names
are used as sample ids.}

\item{outfile}{The name of the mean bigwig file to create.}

\item{aucs}{A vector of equal length to \code{bws} with the AUCs for each
BigWig file. If absent, this infomration will be calculated using
wiggletools.}

\item{chr_sizes}{A single character vector with the path to the chr
sizes file. This depends on the reference genome used for creating the
BigWig files.}

\item{auc_scale}{The value to which the BigWig AUCs will be scaled to. By
default this corresponds to 40 million 100 bp reads.}

\item{wiggletools}{Path to wiggletools. At JHPCE note that you'll have to
use \code{module load wiggletools/default} before opening R.}

\item{wigToBigWig}{Path to wigToBigWig. At JHPCE note that you'll have to use
\code{module load ucsctools} before opening R.}

\item{tempdir}{Path to a temporary directory to use for saving the
intermediate files. You might have to specify one yourself when processing
many BigWig files.}
}
\value{
The path to the mean Wig and BigWig files as well as scripts that 
were created in the process.
}
\description{
Given a set of BigWig files, create a scaled mean BigWig file using the
area under coverage (AUC) to scale the BigWig files. The AUC will be
calculated using wiggletools if necessary. This function requires 
the external dependencies wiggletools and wigToBigWig to work.
}
\details{
Based on parts of 
https://github.com/leekgroup/recount-website/blob/master/recount-prep/prep_merge.R
}
\examples{

if(.Platform$OS.type != 'windows') {
## Disable the example for now. I'd have to figure out how to install
## bwtool on travis
if(FALSE) {
    ## Works at JHPCE. Load the appropriate modules before opening R.
    module load wiggletools/default
    module load ucsctools
    R

    library('recount')
    library('recount.bwtool')
bws <- recount_url$path[match(colData(rse_gene_SRP009615)$bigwig_file, recount_url$file_name)]
    compute_mean(bws[1:2], chr_sizes = '/dcl01/leek/data/gtex_work/runs/gtex/hg38.sizes', tempdir = 'testBW')
}
}

}
\seealso{
\link[recount.bwtool]{coverage_matrix_bwtool}
}
\author{
Leonardo Collado-Torres
}
