% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recount.bwtool-package.R
\docType{package}
\name{recount.bwtool-package}
\alias{recount.bwtool-package}
\title{Compute coverage matrices from recount quickly using bwtool.}
\description{
Compute coverage matrices from recount
\url{https://jhubiostatistics.shinyapps.io/recount/} using \code{bwtool}.
This makes it easy to explore regions beyond the genes and exons that are
available in recount. For example, it can be used for
annotation-agnostic differential expression analyses with the data from the
recount project as described at
\url{http://biorxiv.org/content/early/2016/08/08/068478}.
}
\author{
Leonardo Collado-Torres <leo.collado@libd.org>
}
\keyword{package}
